Refactored sessionStateDemo to conform to geode_examples style
- Excluded gradle files not in root directory from rat
- Added license headers to files which were missing them
