--- sessionState/build.gradle
@@ -1,20 +1,0 @@
-plugins {
-    id 'java'
-    id 'war'
-}
-
-version '1.0-SNAPSHOT'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    // https://mvnrepository.com/artifact/javax.servlet/javax.servlet-api
-    providedCompile group: 'javax.servlet', name: 'javax.servlet-api', version: '3.1.0'
-
-    testCompile group: 'junit', name: 'junit', version: '4.11'
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-}
