+++ settings.gradle
@@ -37,3 +37,4 @@ include 'expiration'
 include 'indexes'
 include 'transaction'
 include 'wan'
+include 'sessionState'
